Refactored the test_launcher (previously test_executor) separating kernel functions from the execution runner.
