+++ test/run_simulation.hpp
@@ -3,65 +3,17 @@
 
 #pragma once
 
-#include <iostream>
-#include <cassert>
-#include <AdePT/BlockData.h>
-inline namespace COPCORE_IMPL {
-
-struct MyTrack {
-  int index{0};
-  float energy{0};
-};
-
-struct Hit {
-  float edep{0};
-};
-
-/** @brief Generate a number of primaries */
-VECCORE_ATT_DEVICE
-void generateAndStorePrimary(int id, adept::BlockData<MyTrack> *tracks)
-{
-  auto track = tracks->NextElement();
-  if (!track) COPCORE_EXCEPTION("generateAndStorePrimary: Not enough space for tracks");
-
-  track->index  = id;
-  track->energy = 100.;
-}
-
-COPCORE_CALLABLE_FUNC(generateAndStorePrimary)
-
-namespace devfunc {
-VECCORE_ATT_DEVICE
-void selectTrack(int id, adept::BlockData<MyTrack> *tracks, adept::mpmc_bounded_queue<int> *queue)
-{
-  auto track    = (*tracks)[id];
-  bool selected = (track.index % 2 == 0);
-  if (selected) queue->enqueue(id);
-}
-COPCORE_CALLABLE_FUNC(selectTrack)
-} // end namespace devfunc
-
-VECCORE_ATT_DEVICE
-void processTrack(MyTrack const &track, Hit &hit)
-{
-  hit.edep += 0.1 * track.energy;
-}
-
-VECCORE_ATT_DEVICE
-void updateTrack(MyTrack &track)
-{
-  track.energy -= 0.1 * track.energy;
-}
+#include "sim_kernels.h"
 
 ///______________________________________________________________________________________
 template <copcore::BackendType backend>
-int simplePipeline()
+int runSimulation()
 {
-  std::cout << "Executing pipeline on " << copcore::BackendName<backend>::name << "\n";
+  std::cout << "Executing simulation on " << copcore::BackendName<backend>::name << "\n";
 
   using TrackBlock     = adept::BlockData<MyTrack>;
   using TrackAllocator = copcore::VariableSizeObjAllocator<TrackBlock, backend>;
-  using HitBlock       = adept::BlockData<Hit>;
+  using HitBlock       = adept::BlockData<MyHit>;
   using HitAllocator   = copcore::VariableSizeObjAllocator<HitBlock, backend>;
   using Queue_t        = adept::mpmc_bounded_queue<int>;
   using QueueAllocator = copcore::VariableSizeObjAllocator<Queue_t, backend>;
@@ -94,16 +46,16 @@ int simplePipeline()
   StreamStruct::CreateStream(stream);
 
   // Allocate some tracks in parallel
-  copcore::Executor<backend> generate(stream);
-  generate.Launch(generateFunc, capacity, {0, 0}, blockT);
+  copcore::Launcher<backend> generate(stream);
+  generate.Run(generateFunc, capacity, {0, 0}, blockT);
 
   // Allow memory to reach the device
   generate.Wait();
 
   std::cout << "Generated " << blockT->GetNused() << " tracks\n";
 
-  copcore::Executor<backend> selector(stream);
-  selector.Launch(selectTrackFunc, blockT->GetNused(), {0, 0}, blockT, queue);
+  copcore::Launcher<backend> selector(stream);
+  selector.Run(selectTrackFunc, blockT->GetNused(), {0, 0}, blockT, queue);
 
   selector.Wait();
   std::cout << "Selected " << queue->size() << " tracks\n";
@@ -118,5 +70,3 @@ int simplePipeline()
   if (!success) return 1;
   return 0;
 }
-
-} // End namespace COPCORE_IMPL
