+++ test/test_launcher.cu
@@ -1,26 +1,18 @@
 // SPDX-FileCopyrightText: 2020 CERN
 // SPDX-License-Identifier: Apache-2.0
+
 /**
  * @file test_track_executor.cu
  * @brief Unit test for the CUDA executor.
  * @author Andrei Gheata (andrei.gheata@cern.ch)
  */
 
-#include <iostream>
-#include <cassert>
-#include <AdePT/BlockData.h>
-#include "executor_functions.h"
-
-int executePipelineGPU();
-int executePipelineCPU()
-{
-  int result = simplePipeline<copcore::BackendType::CPU>();
-  return result;
-}
+#include "run_simulation.hpp"
 
 ///______________________________________________________________________________________
-int main(void)
+int executePipelineGPU()
 {
-  executePipelineCPU();
-  executePipelineGPU();
+  int result;
+  result = runSimulation<copcore::BackendType::CUDA>();
+  return result;
 }
